+++ src/inscriptive/registery_manager/errors/apply_changes_error.rs
@@ -8,7 +8,7 @@ type ContractId = [u8; 32];
 #[derive(Debug, Clone)]
 pub enum RMApplyChangesError {
     AccountTreeOpenError(AccountKey, sled::Error),
-    AccountRegisteryIndexInsertError(AccountKey, u32, sled::Error),
+    AccountRegisteryIndexInsertError(AccountKey, u64, sled::Error),
     AccountCallCounterInsertError(AccountKey, u64, sled::Error),
     AccountBLSKeyInsertError(AccountKey, sled::Error),
     AccountSecondaryAggregationKeyInsertError(AccountKey, sled::Error),
@@ -16,7 +16,7 @@ pub enum RMApplyChangesError {
     AccountNotFoundInMemory(AccountKey),
     AccountCallCounterUpdateError(AccountKey, u64, sled::Error),
     ContractTreeOpenError(ContractId, sled::Error),
-    ContractRegisteryIndexInsertError(ContractId, u32, sled::Error),
+    ContractRegisteryIndexInsertError(ContractId, u64, sled::Error),
     ContractCallCounterInsertError(ContractId, u64, sled::Error),
     ContractProgramBytesInsertError(ContractId, sled::Error),
     ContractNotFoundInMemory(ContractId),
